Fixed race-condition:
The main mutex of the program was left unlocked from launch until the actual main loop. This caused two problems.

One: graphics painting was being done when clients were not ready to start painting.

Two: In very rare, and hard to debug cases, a refresh_request would be made on a client which would gain a lock on the main mutex and start painting it. Then the main thread would catch up and also start painting that client. You'd then have two threads working with cairo and pango at the same time. Pango would check to see if its internal state was valid, find out it's not (as it was being modified by two threads) and crash-assert.

The fix was simply to lock the mutex in app_new, and unlock it in app_main (Where our main loop lives) so that you don't have multiple threads working the same data, at the same time.

Upon looking at request_refresh, it was noticed the amount of threads created could be minimized with an atomic bool on clients signifying that a thread has already launched and will paint that client.