Improved Alsa backend support:
We used to not get the 'Master' volume before, and so we displayed the wrong initial value for what the volume was.

Fixed up battery brightness text display when no suitable brightness controller found.

Also, brought pulseaudio backend in line with how the comments in the pulseaudio library say you're supposed to use it. Although I believe we still sometimes choke on subscribe_cb.

So we no longer deal with threading issues having to do with pulseaudio, I believe we'll have to write something like pa_pollable_mainloop. The synchronous pa_mainloop (which we have to run in a while loop) and the threading issues that happen with pa_threaded_mainloop, are simply not suitable for our program. If pa_mainloop provided a file descriptor which we could poll, and pump events when required, that would be great. But there is sadly no such thing. The good thing is that the source to pa_mainloop is not very long at all.