Keep 'script' search menu always up-to-date.
Before, when a new executable was added to your path, you'd have to open and close the 'script' menu in order to see it listed as an option. Now you can start your search and a thread in the background builds the new script cache and switches it out when it's done, so that you get your expected results without having to manually open and close the menu.