Preview window on thumbnail hover and more...
When you hover over a thumbnail, on KDE (kwin), it will show the full scale preview of that window without changing the stacking order.

Increased visual smoothness when cursor quickly moves from one end of taskbar to the other via testing for intersection of the path from previous mouse position and current.

Taskbar icon breathe animation when launching first window to show that work /is being done/, and the program is just taking a long time to open a window. (or there is something wrong with the launch command)

Fixed README.md