Configure search bar behaviour when hidden
When "Default":

If search field is disabled, will enable it while the app menu is open.

When "Fully Hidden":

If search field is disabled, will still intercept key pressed and start the search menu, but won't display the search field initially.

When "Fully Disabled":

If search field is disabled, will not do anything having to do with search when app menu is open.