Expanded when bounce-zoom animation activates
When a pinned icon has no open apps, and the user clicks to launch the application, a bounce-zoom animation will begin to signify to the user that their click registered and work is being done.

This commit expands that behaviour to when a file is drag-and-dropped onto a pinned icon; and when a user activates a pinned icon using the meta+[0-9] shortcut.