Show "System Settings" search results when on KDE
That is, those KDE settings which are specific to KDE but which users of a desktop taskbar typically expect to be able to search for. Such as display settings, fonts, icons, wallpaper, and so on.

We just had to detect what desktop environment we were on, and then search for the .desktop files which were prefixed with 'kcm_' (the indication that it's a kde search setting).