Ability to specify preferred theme to pull from for icons
In the pinned icon editor, along with the default choices, you'll also see all the theme specific choices available to you.

The default names imply you are letting the system theme be the preferred target theme (which is how it always was), but this update gives you the ability to override the system theme if you want.

You simply prefix the icon with the folder name of where the theme is contained inside two ':', for example: ":Papirus:firefox" will pull the firefox icon from Papirus.