Various changes
Don't do any updates if mouse position is same as it was last frame.

Don't let the pinned icon editor icon dropdown be chosen when pressing enter, if the user hasn't made a move towards preferring drop down options.

Added a Windows 11 Minimal option for date representation in the settings page.

Changed the timing of when the window preview opens so that mousing over when only one window is open takes longer, as it's likely the user simply wanted to switch to that window, rather than looking at the preview.